pyproject.toml
[project]
name = "my-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django>=6.0.1",
    "django-colorfield>=0.14.0",
    "django-widget-tweaks>=1.5.1",
    "gunicorn>=24.1.1",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
]

[dependency-groups]
dev = [
    "honcho>=2.0.0",
    "ruff>=0.14.14",
]
